basic build command refactor
