+++ packages/docusaurus/src/commands/deploy.ts
@@ -12,7 +12,7 @@ import logger from '@docusaurus/logger';
 import shell from 'shelljs';
 import {hasSSHProtocol, buildSshUrl, buildHttpsUrl} from '@docusaurus/utils';
 import {loadContext, type LoadContextParams} from '../server/site';
-import {build} from './build';
+import {build} from './build/build';
 
 export type DeployCLIOptions = Pick<
   LoadContextParams,
